Workshop #2
A program to display the tax brackets for various individuals and households including single, married joint, married separate, and head of household. It will work by getting an income range from the user and displaying the applicable tax for every income with an interval of 1000 within the given field.
The second option of this program will receive the exact amount of the user's taxable income and asks for their filing status. Based on the taxable income and the user's status it will display the applicable tax.